Modified CommandAdminDeleteUser class due to new requirements (PRG pattern)

+++ src/main/java/com/system/command/CommandAdminDeleteUser.java
@@ -1,11 +1,10 @@
 package com.system.command;
 
 import com.system.entity.Account;
-import com.system.entity.User;
+import com.system.manager.HTTPMethod;
 import com.system.manager.ResourceManager;
+import com.system.manager.ServerResponse;
 import com.system.service.AccountService;
-import com.system.service.LetterService;
-import com.system.service.PaymentService;
 import com.system.service.UserService;
 import com.system.utils.Validator;
 
@@ -19,60 +18,52 @@ import java.util.List;
 
 public class CommandAdminDeleteUser implements ICommand {
 
+    // Default path
+    private String pathRedirect = ResourceManager.getInstance().getProperty(ResourceManager.ADMIN_SHOW_USER);
+
     @Override
     public String execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException, SQLException {
 
-        String page = ResourceManager.getInstance().getProperty(ResourceManager.ADMIN_SHOW_USER);
+        String method = request.getMethod();
+        if (request.getMethod().equalsIgnoreCase(HTTPMethod.GET.name())) {
+            return pathRedirect = ResourceManager.getInstance().getProperty(ResourceManager.ADMIN_SHOW_USER);
+        } else if (method.equalsIgnoreCase(HTTPMethod.POST.name())) {
+            pathRedirect = ResourceManager.getInstance().getProperty(ResourceManager.COMMAND_ADMIN_SHOW_USER);
 
-        request.getSession().setAttribute("numberOfLetters", LetterService.getInstance().findUnprocessedLetters().size());
-        request.setAttribute("showUserError", false);
-        request.setAttribute("userHasFundsError", false);
-        request.setAttribute("deleteUserError", false);
-        request.setAttribute("deleted", false);
+            // Data
+            String userIdParam = request.getParameter("userId");
 
-        // Data
-        String userIdParam = request.getParameter("userId");
+            // Validation
+            if (!Validator.checkUserId(userIdParam) || !Validator.checkUserIsAdmin(userIdParam)) {
+                return pathRedirect;
+            }
 
-        // Validation
-        if (!Validator.checkUserId(userIdParam)) {
-            request.setAttribute("deleteUserError", true);
-            return page;
-        }
+            pathRedirect += "&userId=" + userIdParam;
 
-        // Set Attributes
-        Integer userId = Integer.parseInt(userIdParam);
-        setRequestAttributes(request, userId);
+            // Set Attributes
+            Integer userId = Integer.parseInt(userIdParam);
 
-        // Check
-        List<Account> accounts = AccountService.getInstance().findAllAccountsByUserId(userId);
-        for (Account account : accounts) {
-            BigDecimal balance = account.getBalance();
-            if (balance.compareTo(BigDecimal.ZERO) != 0) {
-                request.setAttribute("userHasFundsError", true);
-                return page;
+            // Data
+            List<Account> accounts = AccountService.getInstance().findAllAccountsByUserId(userId);
+            for (Account account : accounts) {
+                BigDecimal balance = account.getBalance();
+                if (balance.compareTo(BigDecimal.ZERO) != 0) {
+                    request.getSession().setAttribute("response", ServerResponse.USER_HAS_FUNDS_ERROR.getResponse());
+                    return pathRedirect;
+                }
+            }
+
+            // Action
+            int status = UserService.getInstance().deleteUserById(userId);
+            if (status == 0) {
+                request.getSession().setAttribute("response", ServerResponse.USER_DELETED_ERROR.getResponse());
+            } else {
+                pathRedirect = ResourceManager.getInstance().getProperty(ResourceManager.COMMAND_INDEX);
+                request.getSession().setAttribute("response", ServerResponse.USER_DELETED_SUCCESS.getResponse());
             }
         }
 
-        // Action
-        int status = UserService.getInstance().deleteUserById(userId);
-        if (status == 0) {
-            request.setAttribute("deleteUserError", true);
-        } else {
-            request.setAttribute("showUserError", true);
-            request.setAttribute("deleted", true);
-        }
-
-        return page;
-    }
-
-    private void setRequestAttributes(HttpServletRequest request, Integer userId) throws SQLException {
-        User user = UserService.getInstance().findUserById(userId);
-        request.setAttribute("userId", userId);
-        request.setAttribute("viewableUser", user);
-        request.setAttribute("paymentsEmpty", PaymentService.getInstance().findLastPaymentsByUserId(userId).isEmpty());
-        request.setAttribute("accountsEmpty", AccountService.getInstance().findAllAccountsByUserId(userId).isEmpty());
-        request.setAttribute("payments", PaymentService.getInstance().findLastPaymentsByUserId(userId));
-        request.setAttribute("accounts", AccountService.getInstance().findAllAccountsByUserId(userId));
+        return pathRedirect;
     }
 
 }
